20250702
Major changes
[1]
Get the list of applicable codecs and store them in Dictionary.
For example, specify H.264.
Extract numbers such as H.264-1 as Keys and store them in SortedDictionary in ascending order.
This is because when getting a list of H.264 in List, it became 3->2->4->1->, and I set it to 3.
[2]
After sorting in ascending order, turn on the live stream of H.264-1.
After that, get the VideoStream with the live stream turned on.
If this does not match, it will be determined that H.264-1 will not be displayed on Genetec.
Move to the next loop of H.264-2.
[3]
The delimiter of the csv has been changed.
This is because "," may be used in the product type.

There may be other minor setting changes. Please feel free to ask if you have any questions.